Registering android fcm token on backend

+++ common/src/commonMain/kotlin/com/marzec/todo/repository/DeviceTokenRepositoryImpl.kt
@@ -6,6 +6,7 @@ import com.marzec.cache.putTyped
 import com.marzec.logger.Logger
 import com.marzec.todo.Api
 import io.ktor.client.HttpClient
+import io.ktor.client.call.body
 import io.ktor.client.request.delete
 import io.ktor.client.request.post
 import io.ktor.client.request.setBody
@@ -16,29 +17,44 @@ class DeviceTokenRepositoryImpl(
     private val client: HttpClient,
     private val logger: Logger,
 ) : DeviceTokenRepository {
-    override suspend fun saveToken(token: String) {
+
+    override suspend fun updateToken(token: String, isLogged: Boolean) {
         try {
             val oldToken = fileCache.getTyped<String>(DEVICE_TOKEN)
             if (oldToken != token) {
-                if (oldToken != null) {
+                if (oldToken != null && isLogged) {
                     client.delete(deleteToken(oldToken))
                 }
                 fileCache.putTyped(DEVICE_TOKEN, token)
-                client.post(FCM_API_URL) {
-                    setBody(CreateFcmTokenDto(token, "android"))
+                if (isLogged) {
+                    client.post(FCM_API_URL) {
+                        setBody(CreateFcmTokenDto(token, "android"))
+                    }.body<Unit>()
                 }
             }
         } catch (e: Exception) {
+            logger.log("DeviceTokenRepositoryImpl", "updateToken", e)
+        }
+    }
+
+    override suspend fun sendCurrentToken() {
+        try {
+            val fcmToken = fileCache.getTyped<String>(DEVICE_TOKEN)
+            fcmToken ?: return
+            val res = client.post(FCM_API_URL) {
+                setBody(CreateFcmTokenDto(fcmToken, "android"))
+            }.body<FcmTokenDto>()
+            logger.log("DeviceTokenRepositoryImpl", res.toString())
+        } catch (e: Exception) {
             logger.log("DeviceTokenRepositoryImpl", "saveToken", e)
         }
-
     }
 
     override suspend fun removeCurrentToken() {
         try {
             val oldToken = fileCache.getTyped<String>(DEVICE_TOKEN)
             oldToken?.let {
-                client.delete(deleteToken(it))
+                client.delete(deleteToken(it)).body<Unit>()
             }
         } catch (e: Exception) {
             logger.log("DeviceTokenRepositoryImpl", "removeCurrentToken", e)
@@ -48,7 +64,7 @@ class DeviceTokenRepositoryImpl(
     private companion object {
         const val DEVICE_TOKEN = "fcm_device_token"
 
-        val FCM_API_URL = "${Api.HOST}/fcm-token"
+        val FCM_API_URL = "${Api.Login.BASE}/fcm-token"
 
         fun deleteToken(token: String) = "$FCM_API_URL/$token"
     }
@@ -58,4 +74,13 @@ class DeviceTokenRepositoryImpl(
 data class CreateFcmTokenDto(
     val fcmToken: String,
     val platform: String? = null
+)
+
+@Serializable
+data class FcmTokenDto(
+    val id: Int,
+    val userId: Int,
+    val fcmToken: String,
+    val platform: String?,
+    val updatedAt: String
 )
